add util module for shared code
