+++ py/desispec/database/spectra.py
@@ -17,6 +17,7 @@ from sqlalchemy import (create_engine, event, ForeignKey, Column, DDL,
 from sqlalchemy.ext.declarative import declarative_base, declared_attr
 from sqlalchemy.orm import scoped_session, sessionmaker, relationship
 from sqlalchemy.schema import CreateSchema
+from .util import convert_dateobs, parse_pgpass
 
 Base = declarative_base()
 engine = None
@@ -402,67 +403,6 @@ def q3c_index(table):
     return
 
 
-def convert_dateobs(timestamp, tzinfo=None):
-    """Convert a string `timestamp` into a :class:`datetime.datetime` object.
-
-    Parameters
-    ----------
-    timestamp : :class:`str`
-        Timestamp in string format.
-    tzinfo : :class:`datetime.tzinfo`, optional
-        If set, add time zone to the timestamp.
-
-    Returns
-    -------
-    :class:`datetime.datetime`
-        The converted `timestamp`.
-    """
-    from datetime import datetime
-    x = datetime.strptime(timestamp, '%Y-%m-%dT%H:%M:%S.%f')
-    if tzinfo is not None:
-        x = x.replace(tzinfo=tzinfo)
-    return x
-
-
-def parse_pgpass(hostname='scidb2.nersc.gov', username='desidev_admin'):
-    """Read a ``~/.pgpass`` file.
-
-    Parameters
-    ----------
-    hostname : :class:`str`, optional
-        Database hostname.
-    username : :class:`str`, optional
-        Database username.
-
-    Returns
-    -------
-    :class:`str`
-        A string suitable for creating a SQLAlchemy database engine, or None
-        if no matching data was found.
-    """
-    from os.path import expanduser
-    fmt = "postgresql://{3}:{4}@{0}:{1}/{2}"
-    try:
-        with open(expanduser('~/.pgpass')) as p:
-            lines = p.readlines()
-    except FileNotFoundError:
-        return None
-    data = dict()
-    for l in lines:
-        d = l.strip().split(':')
-        if d[0] in data:
-            data[d[0]][d[3]] = fmt.format(*d)
-        else:
-            data[d[0]] = {d[3]: fmt.format(*d)}
-    if hostname not in data:
-        return None
-    try:
-        pgpass = data[hostname][username]
-    except KeyError:
-        return None
-    return pgpass
-
-
 def main():
     """Entry point for command-line script.
 
@@ -522,6 +462,9 @@ def main():
     if options.schema:
         schemaname = options.schema
         # event.listen(Base.metadata, 'before_create', CreateSchema(schemaname))
+        if options.clobber:
+            event.listen(Base.metadata, 'before_create',
+                         DDL('DROP SCHEMA IF EXISTS {0} CASCADE'.format(schemaname)))
         event.listen(Base.metadata, 'before_create',
                      DDL('CREATE SCHEMA IF NOT EXISTS {0}'.format(schemaname)))
     #
